# feature(progress): makes performance-log listener show all totals