--- a/test/cli/listeners/performance-log-listener/format-helpers.spec.mjs
+++ b/test/cli/listeners/performance-log-listener/format-helpers.spec.mjs
@@ -20,27 +20,27 @@ import formatHelpers from "../../../../src/cli/listeners/performance-log/format-
 
 describe("[U] cli/listeners/performance-log/format-helpers - formatTime", () => {
   it("converts to ms, left pads & adds the unit at the end", () => {
-    expect(formatHelpers.formatTime(14.88041018)).to.equal("    14,880ms ");
+    expect(formatHelpers.formatTime(14.88041018)).to.equal("     14,880ms ");
   });
 
   it("converts to ms, left pads & adds the unit at the end (0)", () => {
-    expect(formatHelpers.formatTime(0)).to.equal("         0ms ");
+    expect(formatHelpers.formatTime(0)).to.equal("          0ms ");
   });
 
   it("converts to ms, left pads & adds the unit at the end (negative numbers)", () => {
-    expect(formatHelpers.formatTime(-3.1415926535)).to.equal("    -3,142ms ");
+    expect(formatHelpers.formatTime(-3.1415926535)).to.equal("     -3,142ms ");
   });
 
   it("converts to ms, left pads & adds the unit at the end (null treatment => 0)", () => {
-    expect(formatHelpers.formatTime(null)).to.equal("         0ms ");
+    expect(formatHelpers.formatTime(null)).to.equal("          0ms ");
   });
 
   it("converts to ms, left pads & adds the unit at the end (undefined treatment => NaN)", () => {
-    expect(formatHelpers.formatTime()).to.equal("       NaNms ");
+    expect(formatHelpers.formatTime()).to.equal("        NaNms ");
   });
 
   it("converts to ms, left pads & adds the unit at the end (non-number treatment => NaN)", () => {
-    expect(formatHelpers.formatTime("not a number")).to.equal("       NaNms ");
+    expect(formatHelpers.formatTime("not a number")).to.equal("        NaNms ");
   });
 });
 
@@ -56,28 +56,28 @@ describe("[U] cli/listeners/performance-log/format-helpers - formatMemory", () =
   });
 
   it("converts to kB, left pads & adds the unit at the end", () => {
-    expect(formatHelpers.formatMemory(4033856)).to.equal("    +3,939kB ");
+    expect(formatHelpers.formatMemory(4033856)).to.equal("     +3,939kB ");
   });
 
   it("converts to kB, left pads & adds the unit at the end (0)", () => {
-    expect(formatHelpers.formatMemory(0)).to.equal("         0kB ");
+    expect(formatHelpers.formatMemory(0)).to.equal("          0kB ");
   });
 
   it("converts to kB, left pads & adds the unit at the end (negative numbers)", () => {
-    expect(formatHelpers.formatMemory(-403385623)).to.equal("  -393,931kB ");
+    expect(formatHelpers.formatMemory(-403385623)).to.equal("   -393,931kB ");
   });
 
   it("converts to kB, left pads & adds the unit at the end (null)", () => {
-    expect(formatHelpers.formatMemory(0)).to.equal("         0kB ");
+    expect(formatHelpers.formatMemory(0)).to.equal("          0kB ");
   });
 
   it("converts to kB, left pads & adds the unit at the end (undefined)", () => {
-    expect(formatHelpers.formatMemory()).to.equal("       NaNkB ");
+    expect(formatHelpers.formatMemory()).to.equal("        NaNkB ");
   });
 
   it("converts to kB, left pads & adds the unit at the end (not a number)", () => {
     expect(formatHelpers.formatMemory("not a number")).to.equal(
-      "       NaNkB "
+      "        NaNkB "
     );
   });
 });