## Description

- makes the performance-log listener show all totals (instead of only wall-clock time and heap used totals)
- dedicates a separate line to the totals
- widens the width of all columns to accommodate large numbers in locales that use a bit more room.

## Motivation and Context

The listener already showed ∑ wall-clock time and ∑ heap used. This PR adds the ∑ for all other metrics as well (user & system time spent, rss, heap total, external memory). Also instead of putting it in the free format message text it's now in the same column as their per step ∆'s).

The new representation should be easier to read, while at the same time providing more information to analyze performance bottlenecks.


## How Has This Been Tested?

- [x] green ci
- [x] updated automated non-regression test

## Screenshots

A typical output now looks like this (de_DE locale):

```
 elapsed real          user        system         ∆ rss   ∆ heapTotal    ∆ heapUsed    ∆ external after step...
------------- ------------- ------------- ------------- ------------- ------------- ------------- ------------- 
       774 ms        805 ms         97 ms   +132.056 kB    +85.020 kB    +62.670 kB     +2.280 kB start of node process
        12 ms         12 ms          1 ms     +1.144 kB       +256 kB       +785 kB          0 kB parsing options
        80 ms         19 ms          5 ms     +2.536 kB       +548 kB     +1.747 kB       +558 kB cache: check freshness with metadata
         6 ms          6 ms          0 ms     +1.168 kB       +256 kB       +828 kB          0 kB cache: reporting from cache
         1 ms          1 ms          0 ms        +24 kB          0 kB         +5 kB          0 kB really done
------------- ------------- ------------- ------------- ------------- ------------- ------------- ------------- 
       874 ms        842 ms        104 ms   +136.928 kB    +86.080 kB    +66.034 kB     +2.839 kB 
```

Whereas it looked like this before (also de_DE locale):

```
elapsed real         user       system        ∆ rss  ∆ heapTotal   ∆ heapUsed   ∆ external after step...
------------ ------------ ------------ ------------ ------------ ------------ ------------ ------------ 
      770 ms       802 ms        97 ms  +131.624 kB   +85.020 kB   +62.531 kB    +2.280 kB start of node process
       12 ms        11 ms         1 ms    +1.124 kB      +256 kB      +785 kB         0 kB parsing options
       78 ms        18 ms         4 ms    +2.516 kB      +804 kB    +2.010 kB      +558 kB cache: check freshness with metadata
        6 ms         6 ms         0 ms    +1.168 kB         0 kB      +576 kB         0 kB cache: reporting from cache
        1 ms         1 ms         0 ms       +24 kB         0 kB        +5 kB         0 kB really done (867 ms, +65.906 kB)
```

## Types of changes


- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [ ] Refactor (non-breaking change which fixes an issue without changing functionality)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
  - The contribution will be subject to [The MIT license](https://github.com/sverweij/dependency-cruiser/blob/develop/LICENSE), and I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/develop/.github/CONTRIBUTING.md).